midway push (noffin fixed nuffin broke)
Still stuck on get content. implemented document the way index has been implemented so i have SOME access to the data i need, but i'm still not getting it all.
this route might be stupidly wrong but im not getting any guidance on what to do so i've got to try. :)

currently 18:23. spent yesterday on the same thing and got no progress.